feat: VerticalSlider 공통 컴포넌트 구현
